Changed droptofloor, by removing temp vec function and using + operator hurray. However, now sky.cpp is constipated in the ClipSky funcs.

+++ src/svgame/physics.cpp
@@ -819,7 +819,7 @@ void SV_Physics_Step(edict_t *ent)
 
     // friction for flying monsters that have been given vertical velocity
     if ((ent->flags & FL_FLY) && (ent->velocity[2] != 0)) {
-        speed = fabs(ent->velocity[2]);
+        speed = std::fabsf(ent->velocity[2]);
         control = speed < sv_stopspeed ? sv_stopspeed : speed;
         friction = sv_friction / 3;
         newspeed = speed - (FRAMETIME * control * friction);
@@ -831,7 +831,7 @@ void SV_Physics_Step(edict_t *ent)
 
     // friction for flying monsters that have been given vertical velocity
     if ((ent->flags & FL_SWIM) && (ent->velocity[2] != 0)) {
-        speed = fabs(ent->velocity[2]);
+        speed = std::fabsf(ent->velocity[2]);
         control = speed < sv_stopspeed ? sv_stopspeed : speed;
         newspeed = speed - (FRAMETIME * control * sv_waterfriction * ent->waterlevel);
         if (newspeed < 0)
